+++ nomad/acl_endpoint_test.go
@@ -1944,9 +1944,12 @@ func TestACLEndpoint_WhoAmI(t *testing.T) {
 	// Lookup identity claim
 	alloc := mock.Alloc()
 	s1.fsm.State().UpsertAllocs(structs.MsgTypeTestSetup, 1500, []*structs.Allocation{alloc})
-	claims := structs.NewIdentityClaims(alloc.Job, alloc,
+	task := alloc.LookupTask("web")
+	claims := structs.NewIdentityClaimsBuilder(alloc.Job, alloc,
 		wiHandle, // see encrypter_test.go
-		alloc.LookupTask("web").Identity, time.Now().Add(-10*time.Minute))
+		task.Identity).
+		WithTask(task).
+		Build(time.Now().Add(-10 * time.Minute))
 	jwtToken, _, err := s1.encrypter.SignClaims(claims)
 	must.NoError(t, err)
