+++ nomad/alloc_endpoint.go
@@ -619,7 +619,12 @@ func (a *Alloc) signTasks(
 		}
 
 		widFound = true
-		err = a.signIdentities(alloc, wid, idReq, reply, now)
+		claims := structs.NewIdentityClaimsBuilder(alloc.Job, alloc, &idReq.WIHandle, wid).
+			WithTask(task).
+			WithConsul().
+			WithVault().
+			Build(now)
+		err = a.signClaims(claims, idReq, reply)
 		break
 	}
 	return
@@ -638,13 +643,24 @@ func (a *Alloc) signServices(
 	for _, tg := range job.TaskGroups {
 		for _, service := range tg.Services {
 			if service.IdentityHandle(nil).Equal(wid) {
-				return true, a.signIdentities(alloc, service.Identity, idReq, reply, now)
+				claims := structs.NewIdentityClaimsBuilder(
+					alloc.Job, alloc, &idReq.WIHandle, service.Identity).
+					WithConsul().
+					WithService(service).
+					Build(now)
+				return true, a.signClaims(claims, idReq, reply)
 			}
 		}
 		for _, task := range tg.Tasks {
 			for _, service := range task.Services {
 				if service.IdentityHandle(nil).Equal(wid) {
-					return true, a.signIdentities(alloc, service.Identity, idReq, reply, now)
+					claims := structs.NewIdentityClaimsBuilder(
+						alloc.Job, alloc, &idReq.WIHandle, service.Identity).
+						WithTask(task).
+						WithConsul().
+						WithService(service).
+						Build(now)
+					return true, a.signClaims(claims, idReq, reply)
 				}
 			}
 		}
@@ -652,14 +668,11 @@ func (a *Alloc) signServices(
 	return
 }
 
-func (a *Alloc) signIdentities(
-	alloc *structs.Allocation,
-	wid *structs.WorkloadIdentity,
+func (a *Alloc) signClaims(
+	claims *structs.IdentityClaims,
 	idReq *structs.WorkloadIdentityRequest,
 	reply *structs.AllocIdentitiesResponse,
-	now time.Time,
 ) error {
-	claims := structs.NewIdentityClaims(alloc.Job, alloc, &idReq.WIHandle, wid, now)
 	token, _, err := a.srv.encrypter.SignClaims(claims)
 	if err != nil {
 		return err
