+++ nomad/plan_apply.go
@@ -420,7 +420,12 @@ func signAllocIdentities(signer claimSigner, job *structs.Job, allocations []*st
 				continue
 			}
 			defaultWI := &structs.WorkloadIdentity{Name: "default"}
-			claims := structs.NewIdentityClaims(job, alloc, task.IdentityHandle(defaultWI), task.Identity, now)
+
+			claims := structs.NewIdentityClaimsBuilder(
+				job, alloc, task.IdentityHandle(defaultWI), task.Identity).
+				WithTask(task).
+				Build(now)
+
 			token, keyID, err := signer.SignClaims(claims)
 			if err != nil {
 				return err
