+++ nomad/structs/structs.go
@@ -29,7 +29,6 @@ import (
 	"strings"
 	"time"
 
-	jwt "github.com/go-jose/go-jose/v3/jwt"
 	"github.com/hashicorp/cronexpr"
 	"github.com/hashicorp/go-msgpack/v2/codec"
 	"github.com/hashicorp/go-multierror"
@@ -11703,140 +11702,6 @@ func (a *Allocation) LastRescheduleFailed() bool {
 	}
 	return a.RescheduleTracker.LastReschedule != "" &&
 		a.RescheduleTracker.LastReschedule != LastRescheduleSuccess
-}
-
-// IdentityClaims are the input to a JWT identifying a workload. It
-// should never be serialized to msgpack unsigned.
-type IdentityClaims struct {
-	Namespace    string `json:"nomad_namespace"`
-	JobID        string `json:"nomad_job_id"`
-	AllocationID string `json:"nomad_allocation_id"`
-	TaskName     string `json:"nomad_task,omitempty"`
-	ServiceName  string `json:"nomad_service,omitempty"`
-
-	ConsulNamespace string `json:"consul_namespace,omitempty"`
-	VaultNamespace  string `json:"vault_namespace,omitempty"`
-	VaultRole       string `json:"vault_role,omitempty"`
-
-	jwt.Claims
-}
-
-// NewIdentityClaims returns new workload identity claims. Since it may be
-// called with a denormalized Allocation, the Job must be passed in distinctly.
-//
-// ID claim is random (nondeterministic) so multiple calls with the same values
-// will not return equal claims by design. JWT IDs should never collide.
-func NewIdentityClaims(job *Job, alloc *Allocation, wihandle *WIHandle, wid *WorkloadIdentity, now time.Time) *IdentityClaims {
-	tg := job.LookupTaskGroup(alloc.TaskGroup)
-	if tg == nil {
-		return nil
-	}
-
-	if wid == nil {
-		return nil
-	}
-
-	jwtnow := jwt.NewNumericDate(now.UTC())
-	claims := &IdentityClaims{
-		Namespace:    alloc.Namespace,
-		JobID:        alloc.JobID,
-		AllocationID: alloc.ID,
-		Claims: jwt.Claims{
-			NotBefore: jwtnow,
-			IssuedAt:  jwtnow,
-		},
-	}
-
-	// If this is a child job, use the parent's ID
-	if job.ParentID != "" {
-		claims.JobID = job.ParentID
-	}
-
-	var taskName string
-
-	switch wihandle.WorkloadType {
-	case WorkloadTypeService:
-		serviceName := wihandle.WorkloadIdentifier
-		if wihandle.InterpolatedWorkloadIdentifier != "" {
-			serviceName = wihandle.InterpolatedWorkloadIdentifier
-		}
-		claims.ServiceName = serviceName
-
-		// Find task name if this is a task service.
-		for _, t := range tg.Tasks {
-			for _, s := range t.Services {
-				if s.Name == serviceName {
-					taskName = t.Name
-					break
-				}
-			}
-			if taskName != "" {
-				break
-			}
-		}
-
-	case WorkloadTypeTask:
-		taskName = wihandle.WorkloadIdentifier
-		claims.TaskName = taskName
-
-	default:
-		// in case of an unknown workload type we quit
-		return nil
-	}
-
-	// Add ConsulNamespace and VaultNamespace claims if necessary.
-	if taskName != "" {
-		task := tg.LookupTask(taskName)
-		if task == nil {
-			return nil
-		}
-
-		if wid.IsConsul() {
-			if task.Consul != nil {
-				claims.ConsulNamespace = task.Consul.Namespace
-			} else if tg.Consul != nil {
-				claims.ConsulNamespace = tg.Consul.Namespace
-			}
-		}
-
-		if wid.IsVault() && task.Vault != nil {
-			claims.VaultNamespace = task.Vault.Namespace
-			claims.VaultRole = task.Vault.Role
-		}
-
-	} else if wid.IsConsul() && tg.Consul != nil {
-		claims.ConsulNamespace = tg.Consul.Namespace
-	}
-
-	claims.Audience = slices.Clone(wid.Audience)
-	claims.setSubject(job, alloc.TaskGroup, wihandle.WorkloadIdentifier, wid.Name)
-	claims.setExp(now, wid)
-
-	claims.ID = uuid.Generate()
-
-	return claims
-}
-
-// setSubject creates the standard subject claim for workload identities.
-func (claims *IdentityClaims) setSubject(job *Job, group, widentifier, id string) {
-	claims.Subject = strings.Join([]string{
-		job.Region,
-		job.Namespace,
-		job.ID,
-		group,
-		widentifier,
-		id,
-	}, ":")
-}
-
-// setExp sets the absolute time at which these identity claims expire.
-func (claims *IdentityClaims) setExp(now time.Time, wid *WorkloadIdentity) {
-	if wid.TTL == 0 {
-		// No expiry
-		return
-	}
-
-	claims.Expiry = jwt.NewNumericDate(now.Add(wid.TTL))
 }
 
 // AllocationDiff is another named type for Allocation (to use the same fields),
