+++ nomad/structs/workload_id.go
@@ -10,8 +10,10 @@ import (
 	"strings"
 	"time"
 
+	jwt "github.com/go-jose/go-jose/v3/jwt"
 	"github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/go-version"
+	"github.com/hashicorp/nomad/helper/uuid"
 )
 
 const (
@@ -60,6 +62,168 @@ var (
 	// Nomad 1.7.0 as well
 	MinNomadVersionVaultWID = version.Must(version.NewVersion("1.7.0-a"))
 )
+
+// IdentityClaims are the input to a JWT identifying a workload. It
+// should never be serialized to msgpack unsigned.
+type IdentityClaims struct {
+	Namespace    string `json:"nomad_namespace"`
+	JobID        string `json:"nomad_job_id"`
+	AllocationID string `json:"nomad_allocation_id"`
+	TaskName     string `json:"nomad_task,omitempty"`
+	ServiceName  string `json:"nomad_service,omitempty"`
+
+	ConsulNamespace string `json:"consul_namespace,omitempty"`
+	VaultNamespace  string `json:"vault_namespace,omitempty"`
+	VaultRole       string `json:"vault_role,omitempty"`
+
+	jwt.Claims
+}
+
+// IdentityClaimsBuilder is used to build up all the context we need to create
+// IdentityClaims from jobs, allocs, tasks, services, Vault and Consul
+// configurations, etc. This lets us treat IdentityClaims as the immutable
+// output of that process.
+type IdentityClaimsBuilder struct {
+	wid         *WorkloadIdentity // from jobspec
+	wihandle    *WIHandle
+	alloc       *Allocation
+	job         *Job
+	tg          *TaskGroup
+	task        *Task
+	serviceName string
+	consul      *Consul
+	vault       *Vault
+}
+
+// NewIdentityClaimsBuilder returns an initialized IdentityClaimsBuilder for the
+// allocation and identity request. Because it may be called with a denormalized
+// Allocation in the plan applier, the Job must be passed in as a separate
+// parameter.
+func NewIdentityClaimsBuilder(job *Job, alloc *Allocation, wihandle *WIHandle, wid *WorkloadIdentity) *IdentityClaimsBuilder {
+	tg := job.LookupTaskGroup(alloc.TaskGroup)
+	if tg == nil {
+		return nil
+	}
+
+	return &IdentityClaimsBuilder{
+		alloc:    alloc,
+		job:      job,
+		wihandle: wihandle,
+		wid:      wid,
+		tg:       tg,
+	}
+}
+
+// WithTask adds a task to the builder context.
+func (b *IdentityClaimsBuilder) WithTask(task *Task) *IdentityClaimsBuilder {
+	if task == nil {
+		return b
+	}
+	b.task = task
+	return b
+}
+
+// WithVault adds the task's vault block to the builder context. This should
+// only be called after WithTask.
+func (b *IdentityClaimsBuilder) WithVault() *IdentityClaimsBuilder {
+	if !b.wid.IsVault() || b.task == nil {
+		return b
+	}
+	b.vault = b.task.Vault
+	return b
+}
+
+// WithConsul adds the group or task's consul block to the builder context. For
+// task identities, this should only be called after WithTask.
+func (b *IdentityClaimsBuilder) WithConsul() *IdentityClaimsBuilder {
+	if !b.wid.IsConsul() {
+		return b
+	}
+	if b.task != nil && b.task.Consul != nil {
+		b.consul = b.task.Consul
+	} else if b.tg.Consul != nil {
+		b.consul = b.tg.Consul
+	}
+	return b
+}
+
+// WithService adds a service block to the builder context. This should only be
+// called for service identities, and a builder for service identities will
+// never set the task_name claim.
+func (b *IdentityClaimsBuilder) WithService(service *Service) *IdentityClaimsBuilder {
+	if b.wihandle.WorkloadType != WorkloadTypeService {
+		return b
+	}
+	serviceName := b.wihandle.WorkloadIdentifier
+	if b.wihandle.InterpolatedWorkloadIdentifier != "" {
+		serviceName = b.wihandle.InterpolatedWorkloadIdentifier
+	}
+	b.serviceName = serviceName
+	return b
+}
+
+// Build is the terminal method for the builder and sets all the derived values
+// on the claim. The claim ID is random (nondeterministic) so multiple calls
+// with the same values will not return equal claims by design. JWT IDs should
+// never collide.
+func (b *IdentityClaimsBuilder) Build(now time.Time) *IdentityClaims {
+
+	jwtnow := jwt.NewNumericDate(now.UTC())
+	claims := &IdentityClaims{
+		Namespace:    b.alloc.Namespace,
+		JobID:        b.job.ID,
+		AllocationID: b.alloc.ID,
+		ServiceName:  b.serviceName,
+		Claims: jwt.Claims{
+			NotBefore: jwtnow,
+			IssuedAt:  jwtnow,
+		},
+	}
+	// If this is a child job, use the parent's ID
+	if b.job.ParentID != "" {
+		claims.JobID = b.job.ParentID
+	}
+	if b.task != nil && b.wihandle.WorkloadType != WorkloadTypeService {
+		claims.TaskName = b.task.Name
+	}
+	if b.consul != nil {
+		claims.ConsulNamespace = b.consul.Namespace
+	}
+	if b.vault != nil {
+		claims.VaultNamespace = b.vault.Namespace
+		claims.VaultRole = b.vault.Role
+	}
+
+	claims.Audience = slices.Clone(b.wid.Audience)
+	claims.setSubject(b.job, b.alloc.TaskGroup, b.wihandle.WorkloadIdentifier, b.wid.Name)
+	claims.setExp(now, b.wid)
+
+	claims.ID = uuid.Generate()
+
+	return claims
+}
+
+// setSubject creates the standard subject claim for workload identities.
+func (claims *IdentityClaims) setSubject(job *Job, group, widentifier, id string) {
+	claims.Subject = strings.Join([]string{
+		job.Region,
+		job.Namespace,
+		job.ID,
+		group,
+		widentifier,
+		id,
+	}, ":")
+}
+
+// setExp sets the absolute time at which these identity claims expire.
+func (claims *IdentityClaims) setExp(now time.Time, wid *WorkloadIdentity) {
+	if wid.TTL == 0 {
+		// No expiry
+		return
+	}
+
+	claims.Expiry = jwt.NewNumericDate(now.Add(wid.TTL))
+}
 
 // WorkloadIdentity is the jobspec block which determines if and how a workload
 // identity is exposed to tasks similar to the Vault block.
